move synthetic enums out of metamodels
