+++ .changeset/selfish-walls-exercise.md
@@ -1,0 +1,5 @@
+---
+'@finos/legend-graph': major
+---
+
+**BREAKING CHANGE:** Move some `enum`s out of metamodels, such as `SET_IMPLEMENTATION_TYPE`, `BASIC_SET_IMPLEMENTATION_TYPE`, `PACKAGEABLE_ELEMENT_POINTER_TYPE`, `CLASS_PROPERTY_TYPE`, etc.
