+++ packages/legend-graph/src/models/metamodels/pure/packageableElements/mapping/SetImplementation.ts
@@ -105,21 +105,3 @@ export abstract class SetImplementation
     visitor: SetImplementationVisitor<T>,
   ): T;
 }
-
-// TODO: to be moved out of metamodel
-export enum BASIC_SET_IMPLEMENTATION_TYPE {
-  OPERATION = 'operation',
-  INSTANCE = 'instance',
-}
-
-// TODO: to be moved out of metamodel
-export enum SET_IMPLEMENTATION_TYPE {
-  OPERATION = 'operation',
-  MERGE_OPERATION = 'mergeOperation',
-  PUREINSTANCE = 'pureInstance',
-  FLAT_DATA = 'flatData',
-  EMBEDDED_FLAT_DATA = 'embeddedFlatData',
-  RELATIONAL = 'relational',
-  EMBEDDED_RELATIONAL = 'embeddedRelational',
-  AGGREGATION_AWARE = 'aggregationAware',
-}
