+++ packages/legend-studio/src/components/editor/edit-panel/mapping-editor/PurePropertyMappingEditor.tsx
@@ -38,8 +38,6 @@ import { type ConnectDropTarget, useDrop } from 'react-dnd';
 import { useEditorStore } from '../../EditorStoreProvider';
 import {
   Enumeration,
-  CLASS_PROPERTY_TYPE,
-  getClassPropertyType,
   EnumerationMapping,
   DerivedProperty,
   getEnumerationMappingsByEnumeration,
@@ -47,6 +45,10 @@ import {
 import { StudioLambdaEditor } from '../../../shared/StudioLambdaEditor';
 import { purePropertyMapping_setTransformer } from '../../../../stores/graphModifier/DSLMapping_GraphModifierHelper';
 import { getExpectedReturnType } from './PropertyMappingsEditor';
+import {
+  CLASS_PROPERTY_TYPE,
+  getClassPropertyType,
+} from '../../../../stores/shared/ModelUtil';
 
 const SimplePropertyMappingEditor = observer(
   (props: {
